feat(CLI): Extracted CLI functions to its own class for cleaner syntax
