+++ lib/bdd/index.js
@@ -13,8 +13,7 @@ const { updateSettings } = require('./core/utils');
  * @param {object} settings The settings override
  * @param {object} customData Data added by user to scope object
  */
-const Cruller = function(settings = {}, customData = {}) {
-    
+const CrullerCucumber = function(settings = {}, customData = {}) {
     const cucumberWorld = function({ attach, parameters }) {
         // The settings object
         updateSettings(settings);
@@ -42,7 +41,7 @@ const Cruller = function(settings = {}, customData = {}) {
     setWorldConstructor(cucumberWorld);
 };
 
-module.exports = Cruller;
+module.exports = CrullerCucumber;
 module.exports.scope = scope;
 module.exports.actions = actions;
 module.exports.steps = steps;
